add benchmarks and tests

+++ hw01algebraicAlgorithms/src/main/java/com/alextim/Exponentiation.java
@@ -2,15 +2,15 @@ package com.alextim;
 
 public class Exponentiation {
 
-    public int simple(int base, int power) {
-        int a = 1;
+    public double simple(double base, int power) {
+        double a = 1;
         for (int i = 0; i<power; i++)
             a *= base;
         return a;
     }
 
-    public int withHelpPowerOf2(int base, int power) {
-        final int base0 = base;
+    public double withHelpPowerOf2(double base, int power) {
+        final double base0 = base;
         int powerOf2 = (int)(Math.log(power) / Math.log(2));
 
         for (int i = 0; i<powerOf2; i++)
@@ -24,8 +24,8 @@ public class Exponentiation {
         return base;
     }
 
-    public int withHelpBinaryDecompositionOfPower(int base, int power) {
-        int res = 1;
+    public double withHelpBinaryDecompositionOfPower(double base, int power) {
+        double res = 1;
         while (power != 0) {
             if(power % 2 != 0) {
                 res *= base;
